Added ExternalAsset
