+++ dashx/src/main/java/com/dashx/sdk/data/ExternalAsset.kt
@@ -1,0 +1,18 @@
+package com.dashx.sdk.data
+
+data class ExternalAsset(
+    val id: String,
+    val externalColumnId: String,
+    val status: String,
+    val data: ExternalAssetData,
+)
+
+data class ExternalAssetData(
+    val asset: AssetData,
+    val upload: AssetData
+)
+
+data class AssetData(
+    val status: String,
+    val url: String
+)
